Implements connection logic.
Connection logic is implemented and abstracted in class
WLXBluetoothConnectionManager.

In case of connection lost, retry is implemented using
a reconnection strategy. The only implemented strategy
for the moment is linear backoff, implemented in
WLXLinearReconnectionStrategy. Other strategies can
be implemented and injected by implementing the protocol
WLXReconnectionStrategy.

All the glue code between the discovery process and the
connection process was placed in WLXBluetoothDeviceManager.
Which also persist the last connected device using a repository
so it then a user can connect with the last connected device
without the need of discovery.

The repository is abstracted in a protocol called
WLXBluetoothDeviceRepository and the only implementation
for the moment uses NSUserDefaults and is in
WLXBluetoothDeviceUserDefaultsRepository.