Extracts characteristic async logic.
Executing a block of code for a characteristic given its UUID
has been extracted to WLXCharacteristicAsyncExecutor. This class
is in charge of executing a block of code for a characteristic
in a separate dispatch queue. If the characteristic has not been
discovered, this class uses the service manager to discover the
characteristic and one the characteristic is available the
block of code gets executed.

Also a common structure, used to save handlers/observer of
characteristic, a dictionary of array has been extracted to
WLXDictionaryOfArray.