Invalidates the services and all the service managers. #27
When a connection is lost, terminated or the app is
reconnecting the services manager and all of its service
manager are invalidated. This means that calling
any method on an invalidated object will raise an
exception.

The after a reconnectio the user should obtain
a new services manager and new services manager.